COMMERCE-875 commerce-openapi-util - Extract Provider to use it easily in other classes
